svg/text/parsing/shape-inside-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Text: parsing shape-inside with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/text.html#ShapeInsideProperty" rel="help"/><h:meta content="shape-inside supports only the grammar 'auto | [ basic-shape | uri ]+'" name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("shape-inside", "none");
test_invalid_value("shape-inside", "auto circle(10px at 20px -30px)");
test_invalid_value("shape-inside", 'url("https://example.com/") auto');

test_invalid_value("shape-inside", "inset(25%)");

  ]]></script></svg>
